requirements.txt
asgiref       3.2.10
Django        3.1
django-filter 21.1
pip           22.0.4
pytz          2021.3
setuptools    60.10.0
sqlparse      0.4.2
tzdata        2021.5

